Split tests with BaseClass
