+++ .eslintrc.js
@@ -351,7 +351,8 @@ module.exports = {
             },
             "rules": {
                 "global-require": 0,
-                "import/no-dynamic-require": 0
+                "import/no-dynamic-require": 0,
+                "no-new": 0
             }
         }
     ]
