+++ src/__tests__/webpack/index.test.ts
@@ -5,24 +5,7 @@
 import { BuildPlugin } from '../../webpack';
 import { mockCompiler } from '../helpers/testHelpers';
 
-type ConsoleError = (message?: any, ...optionalParams: any[]) => void;
-
 describe('webpack', () => {
-    test('It should initialise correctly', () => {
-        const plugin = new BuildPlugin();
-        expect(plugin.name).toBe('BuildPlugin');
-        expect(plugin.hooks.length).toBe(3);
-    });
-
-    test('It should register custom hooks', () => {
-        const plugin = new BuildPlugin({
-            hooks: ['./src/__tests__/mocks/customHook.ts'],
-        });
-
-        expect(plugin.hooks.length).toBe(4);
-        expect(typeof plugin.hooks[3].hooks.preoutput).toBe('function');
-    });
-
     test(`It should use given context or default to webpack's configuration`, () => {
         const plugin1 = new BuildPlugin({
             context: '/fake/path',
@@ -38,21 +21,6 @@ describe('webpack', () => {
 
         expect(plugin1.options.context).toBe('/fake/path');
         expect(plugin2.options.context).toBe('/default/context');
-    });
-
-    test('It should log, given a broken hook path', () => {
-        console.error = (jest
-            .spyOn(console, 'error')
-            .mockImplementation(
-                (message?: any, ...optionalParams: any[]) => {}
-            ) as unknown) as ConsoleError;
-
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const plugin = new BuildPlugin({
-            hooks: ['./broken/path.ts'],
-        });
-
-        expect(console.error).toHaveBeenCalledTimes(1);
     });
 
     test('It should not execute if disabled', () => {
